Do not wrap queue size around UINT_MAX

+++ src/shared_queue.hpp
@@ -70,7 +70,7 @@ public:
       return queue_.empty();
    }
 
-   unsigned size() const {
+   size_t size() const {
       std::lock_guard<std::mutex> lock(m_);
       return queue_.size();
    }
